requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
uvicorn==0.13.2
fastapi==0.63.0
pydantic==1.8.1
starlette==0.13.6
typing-extensions==3.7.4.3
greenlet==1.0.0
sqlalchemy==1.4.10
python-dotenv==0.17.0
psycopg2-binary==2.8.6
cffi==1.14.5
cryptography==3.4.7
pyjwt==2.0.1
pycparser==2.20
cryptography==3.4.7
pycparser==2.20
passlib==1.7.4
